update lobby participant count when participants leave

+++ express/socket.js
@@ -7,7 +7,7 @@ module.exports = function (io) {
         let currentRoomNumber;
 
         socket.on('room', function (roomNumber) {
-            console.log("User " + socket.id + " wants to join room " + roomNumber);
+            console.log("User " + socket.id + " is attempting to join room " + roomNumber);
             socket.join(roomNumber); // join client into room roomNumber
             currentRoomNumber = roomNumber;
 
@@ -28,29 +28,33 @@ module.exports = function (io) {
         });
 
         socket.on("disconnect", () => {
-            // Update participants in room when they leave
-            let query = { roomNumber: currentRoomNumber };
-            Room.findOneAndUpdate(query,
-                { $pull: { participants: { socketID: socket.id } } },
-                { new: true }).then(result => {
-                    if (result) {
-                        // Check if last participant has disconnected
-                        if (result.participants.length === 0) {
-                            Room.findOneAndDelete(query).then(result => {
-                                if (!result) {
-                                    console.log("Failed to delete room.");
-                                } else {
-                                    console.log("Deleted room", currentRoomNumber);
-                                }
-                            });
+            if (currentRoomNumber) {
+                // Update participants in room when they leave
+                let query = { roomNumber: currentRoomNumber };
+                Room.findOneAndUpdate(query,
+                    { $pull: { participants: { socketID: socket.id } } },
+                    { new: true }).then(result => {
+                        if (result) {
+                            // Check if last participant has disconnected
+                            if (result.participants.length === 0) {
+                                Room.findOneAndDelete(query).then(result => {
+                                    if (!result) {
+                                        console.log("Failed to delete room.");
+                                    } else {
+                                        console.log("Deleted room", currentRoomNumber);
+                                    }
+                                });
+                            }
+                            roomInfo = JSON.stringify(result);
+                            io.sockets.in(currentRoomNumber).emit('user disconnect', roomInfo);
+                            console.log("User " + socket.id + " has disconnected from room " + currentRoomNumber);
+                        } else {
+                            console.log("Room to disconnect from does not exist."); // TODO: handle properly
                         }
-                    } else {
-                        console.log("Room to disconnect from does not exist."); // TODO: handle properly
-                    }
-                }).catch(e => console.log(e));
-            console.log("User " + socket.id + " has disconnected from room " + currentRoomNumber);
-            let message = "User " + " has left your room.";
-            io.sockets.in(currentRoomNumber).emit('user disconnect', message);
+                    }).catch(e => console.log(e));
+            } else {
+                console.log("User" + socket.id + "disconnected without having ever joined a room.");
+            }
         });
     });
 }
